// cosmowcl/src/app/shared/models/api/cosmo/completeUploadDocumentiZipSessionRequest.d.ts
/*
 * Copyright CSI-Piemonte - 2023
 * SPDX-License-Identifier: GPL-3.0-or-later
 */
/**
 * session.v2
 * Servizi per la gestione degli upload
 *
 * The version of the OpenAPI document: 2.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


export interface CompleteUploadDocumentiZipSessionRequest { 
    hashes?: string;
    folderName: string;
}